EPMRPP-109656 events refactoring
