--- src/main/java/com/epam/reportportal/extension/azure/event/handler/EventHandler.java
@@ -1,9 +1,0 @@
-package com.epam.reportportal.extension.azure.event.handler;
-
-/**
- * @author <a href="mailto:ivan_budayeu@epam.com">Ivan Budayeu</a>
- */
-public interface EventHandler<T> {
-
-	void handle(T event);
-}
