--- src/main/java/com/epam/reportportal/extension/azure/event/handler/plugin/PluginLoadedEventHandler.java
@@ -1,53 +1,0 @@
-package com.epam.reportportal.extension.azure.event.handler.plugin;
-
-import com.epam.reportportal.extension.azure.event.handler.EventHandler;
-import com.epam.reportportal.extension.azure.info.PluginInfoProvider;
-import com.epam.reportportal.extension.event.PluginEvent;
-import com.epam.reportportal.infrastructure.persistence.dao.IntegrationRepository;
-import com.epam.reportportal.infrastructure.persistence.dao.IntegrationTypeRepository;
-import com.epam.reportportal.infrastructure.persistence.entity.integration.Integration;
-import com.epam.reportportal.infrastructure.persistence.entity.integration.IntegrationParams;
-import com.epam.reportportal.infrastructure.persistence.entity.integration.IntegrationType;
-import java.time.Instant;
-import java.util.HashMap;
-import java.util.List;
-
-/**
- * @author <a href="mailto:ivan_budayeu@epam.com">Ivan Budayeu</a>
- */
-public class PluginLoadedEventHandler implements EventHandler<PluginEvent> {
-
-	private final IntegrationTypeRepository integrationTypeRepository;
-	private final IntegrationRepository integrationRepository;
-	private final PluginInfoProvider pluginInfoProvider;
-
-	public PluginLoadedEventHandler(IntegrationTypeRepository integrationTypeRepository, IntegrationRepository integrationRepository,
-			PluginInfoProvider pluginInfoProvider) {
-		this.integrationTypeRepository = integrationTypeRepository;
-		this.integrationRepository = integrationRepository;
-		this.pluginInfoProvider = pluginInfoProvider;
-	}
-
-	@Override
-	public void handle(PluginEvent event) {
-		integrationTypeRepository.findByName(event.getPluginId()).ifPresent(integrationType -> {
-			createIntegration(event.getPluginId(), integrationType);
-			integrationTypeRepository.save(pluginInfoProvider.provide(integrationType));
-		});
-	}
-
-	private void createIntegration(String name, IntegrationType integrationType) {
-		List<Integration> integrations = integrationRepository.findAllGlobalByType(integrationType);
-		if (integrations.isEmpty()) {
-			Integration integration = new Integration();
-			integration.setName(name);
-			integration.setType(integrationType);
-			integration.setCreationDate(Instant.now());
-			integration.setEnabled(true);
-			integration.setCreator("SYSTEM");
-			integration.setParams(new IntegrationParams(new HashMap<>()));
-			integrationRepository.save(integration);
-		}
-	}
-
-}
