--- src/main/java/com/epam/reportportal/extension/azure/event/plugin/AzurePluginEventListener.java
@@ -1,33 +1,0 @@
-package com.epam.reportportal.extension.azure.event.plugin;
-
-import static java.util.Optional.ofNullable;
-
-import com.epam.reportportal.extension.azure.event.EventHandlerFactory;
-import com.epam.reportportal.extension.event.PluginEvent;
-import org.springframework.context.ApplicationListener;
-
-/**
- * @author <a href="mailto:ivan_budayeu@epam.com">Ivan Budayeu</a>
- */
-public class AzurePluginEventListener implements ApplicationListener<PluginEvent> {
-
-	private final String pluginId;
-	private final EventHandlerFactory<PluginEvent> pluginEventEventHandlerFactory;
-
-	public AzurePluginEventListener(String pluginId, EventHandlerFactory<PluginEvent> pluginEventEventHandlerFactory) {
-		this.pluginId = pluginId;
-		this.pluginEventEventHandlerFactory = pluginEventEventHandlerFactory;
-	}
-
-	@Override
-	public void onApplicationEvent(PluginEvent event) {
-		if (supports(event)) {
-			ofNullable(pluginEventEventHandlerFactory.getEventHandler(event.getType())).ifPresent(pluginEventEventHandler -> pluginEventEventHandler
-					.handle(event));
-		}
-	}
-
-	private boolean supports(PluginEvent event) {
-		return pluginId.equals(event.getPluginId());
-	}
-}
