--- src/main/java/com/epam/reportportal/extension/azure/event/plugin/PluginEventHandlerFactory.java
@@ -1,34 +1,0 @@
-package com.epam.reportportal.extension.azure.event.plugin;
-
-import com.epam.reportportal.extension.azure.event.EventHandlerFactory;
-import com.epam.reportportal.extension.azure.event.handler.EventHandler;
-import com.epam.reportportal.extension.azure.event.handler.plugin.PluginLoadedEventHandler;
-import com.epam.reportportal.extension.azure.info.PluginInfoProvider;
-import com.epam.reportportal.extension.event.PluginEvent;
-import com.epam.reportportal.infrastructure.persistence.dao.IntegrationRepository;
-import com.epam.reportportal.infrastructure.persistence.dao.IntegrationTypeRepository;
-import java.util.HashMap;
-import java.util.Map;
-
-/**
- * @author <a href="mailto:ivan_budayeu@epam.com">Ivan Budayeu</a>
- */
-public class PluginEventHandlerFactory implements EventHandlerFactory<PluginEvent> {
-
-	public static final String LOAD_KEY = "load";
-
-	private final Map<String, EventHandler<PluginEvent>> eventHandlerMapping;
-
-	public PluginEventHandlerFactory(IntegrationTypeRepository integrationTypeRepository, IntegrationRepository integrationRepository,
-			PluginInfoProvider pluginInfoProvider) {
-		this.eventHandlerMapping = new HashMap<>();
-		this.eventHandlerMapping.put(LOAD_KEY,
-				new PluginLoadedEventHandler(integrationTypeRepository, integrationRepository, pluginInfoProvider)
-		);
-	}
-
-	@Override
-	public EventHandler<PluginEvent> getEventHandler(String key) {
-		return eventHandlerMapping.get(key);
-	}
-}
